Former-commit-id: c79633a38927fe4039a0dd4c4b6e6f12d484d3ba

+++ sys/dev/isp/isp_freebsd.h
@@ -1823,7 +1823,7 @@ name|XS_LUN
 parameter_list|(
 name|ccb
 parameter_list|)
-value|(ccb)->ccb_h.target_lun
+value|(uint32_t)((ccb)->ccb_h.target_lun)
 end_define
 
 begin_define
